pyproject.toml
[project]
name = "first-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio>=5.20.0",
    "matplotlib>=3.10.1",
    "openinference-instrumentation-smolagents>=0.1.6",
    "opentelemetry-exporter-otlp>=1.30.0",
    "opentelemetry-sdk>=1.30.0",
    "pandas>=2.2.3",
    "smolagents[openai]>=1.9.2",
]
